Added option to change the Request type
You can now choose if you want the message to be streamed or not. I find that lately the streaming has been spotty, just sending plain request seems to be more consistent even though its not as pretty.

This is not done as ATM there is a token limit unlike in the streaming mode.